// src/vendor/types.ts
export interface localChar {
	name: string;
	info: {
		gs: number;
		clas: string;
		id: number;
		isChecked: {
			chaos: boolean;
			guards: boolean;
			daily: boolean;
			w_guards: boolean;
			argos: boolean;
			boldan: boolean;
			bela: boolean;
			clown: boolean;
			avr: boolean;
		};
	};
}
export interface ICheckbox {
	chaos: boolean;
	guards: boolean;
	daily: boolean;
	w_guards: boolean;
	argos: boolean;
	boldan: boolean;
	bela: boolean;
	clown: boolean;
	avr: boolean;
}